refactor(auth): reorganize auth modules and unify response format
