+++ backend/src/auth/domain.rs
@@ -1,0 +1,29 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct Claims {
+    pub sub: String,
+    pub exp: usize,
+}
+
+#[derive(Debug, Deserialize)]
+pub struct SignupRequest {
+    pub login_name: String,
+    pub password: String,
+}
+
+#[derive(Debug, Serialize)]
+pub struct SignupResponse {
+    pub login_name: String,
+}
+
+#[derive(Debug, Deserialize)]
+pub struct LoginRequest {
+    pub login_name: String,
+    pub password: String,
+}
+
+#[derive(Debug, Serialize)]
+pub struct LoginResponse {
+    pub token: String,
+}
