+++ backend/src/auth/mod.rs
@@ -1,0 +1,5 @@
+pub mod domain;
+pub mod jwt;
+pub mod password;
+pub mod route;
+pub mod service;
