+++ backend/src/auth/service/mod.rs
@@ -1,0 +1,5 @@
+pub mod login;
+pub mod signup;
+
+pub use login::login;
+pub use signup::signup;
