+++ backend/src/routes/mod.rs
@@ -1,2 +1,1 @@
-pub mod auth;
 pub mod parts;
