+++ backend/src/services/mod.rs
@@ -1,2 +1,1 @@
-// pub mod part_service;
 pub mod part;
